Music & visualizer tweaks
Changed all music to be 2D rather than 3D (so we don't need to worry
about where the audiolistener is relative to the source) and cut out the
first 16 measures of Lunar for the menu when it loads via the splash
(starts playing during the splash screen).  Also started working on more
dynamic spectrum analyzers.  They can handle position and rotation now,
as well as having multiple in one scene, but I'm still working on
dynamic width.  I put one on the left of Colors just as an example/test.

Also, we can move the whole thing as a unit, allowing for animations
across the scene or whatever.  Maybe we can also use it for the menu
buttons?  idk